+++ Source/Plugin.BLE/Android/BroadcastReceivers/BondStatusBroadcastReceiver.cs
@@ -1,41 +1,54 @@
 using System;
 using Android.Bluetooth;
 using Android.Content;
-using Plugin.BLE.Abstractions;
+using Android.OS;
 using Plugin.BLE.Abstractions.EventArgs;
 using Plugin.BLE.Android;
+using Plugin.BLE.Extensions;
 
 namespace Plugin.BLE.BroadcastReceivers
 {
-    //[BroadcastReceiver]
     public class BondStatusBroadcastReceiver : BroadcastReceiver
     {
+	    private readonly Adapter _broadcastAdapter;
+
         public event EventHandler<DeviceBondStateChangedEventArgs> BondStateChanged;
+
+        public BondStatusBroadcastReceiver(Adapter adapter)
+        {
+            _broadcastAdapter = adapter;
+        }
 
         public override void OnReceive(Context context, Intent intent)
         {
-            var bondState = (Bond)intent.GetIntExtra(BluetoothDevice.ExtraBondState, (int)Bond.None);
-            //ToDo
-            var device = new Device(null, (BluetoothDevice)intent.GetParcelableExtra(BluetoothDevice.ExtraDevice), null);
-            Console.WriteLine(bondState.ToString());
-
-            if (BondStateChanged == null) return;
-
-            switch (bondState)
+            if (BondStateChanged == null)
             {
-                case Bond.None:
-                    BondStateChanged(this, new DeviceBondStateChangedEventArgs() { Device = device, State = DeviceBondState.NotBonded });
-                    break;
-
-                case Bond.Bonding:
-                    BondStateChanged(this, new DeviceBondStateChangedEventArgs() { Device = device, State = DeviceBondState.Bonding });
-                    break;
-
-                case Bond.Bonded:
-                    BondStateChanged(this, new DeviceBondStateChangedEventArgs() { Device = device, State = DeviceBondState.Bonded });
-                    break;
-
+	            return;
             }
+
+            var extraBondState = (Bond)intent.GetIntExtra(BluetoothDevice.ExtraBondState, (int)Bond.None);
+
+            BluetoothDevice bluetoothDevice;
+            
+#if NET6_0_OR_GREATER
+            if (OperatingSystem.IsAndroidVersionAtLeast(33))
+#else
+            if (Build.VERSION.SdkInt >= BuildVersionCodes.Tiramisu)
+#endif
+            {
+	            bluetoothDevice = (BluetoothDevice)intent.GetParcelableExtra(BluetoothDevice.ExtraDevice, Java.Lang.Class.FromType(typeof(BluetoothDevice)));
+            }
+            else 
+            {
+	            bluetoothDevice = (BluetoothDevice)intent.GetParcelableExtra(BluetoothDevice.ExtraDevice);
+            }
+            
+            var device = new Device(_broadcastAdapter, bluetoothDevice, null);
+
+            var address = bluetoothDevice?.Address ?? string.Empty;
+
+            var bondState = extraBondState.FromNative();
+            BondStateChanged(this, new DeviceBondStateChangedEventArgs() { Address = address, Device = device, State = bondState });
         }
     }
 }